Improve test runtime
Squashed commit of the following:

commit 88404fd22a81384eba0721082fbb5e703d9b4810
Author: Eugenio Lacuesta <eugenio.lacuesta@gmail.com>
Date:   Sat Jun 13 12:35:20 2026 -0300

    Reduce timeout for disabled JS

commit 9973204b9db94aa5513670a1e0c048b801cb05fe
Author: Eugenio Lacuesta <eugenio.lacuesta@gmail.com>
Date:   Sat Jun 13 12:27:31 2026 -0300

    Remove duplicated request

commit c80b44717e201ec20a6d4bc1abeb6cbaa2f44c95
Author: Eugenio Lacuesta <eugenio.lacuesta@gmail.com>
Date:   Sat Jun 13 12:20:15 2026 -0300

    Poll TCP port for browser readiness

commit 7ee95e7240f6cd8a40f16d58983aa755048bd09f
Author: Eugenio Lacuesta <eugenio.lacuesta@gmail.com>
Date:   Sat Jun 13 12:17:40 2026 -0300

    Reduce delays in tests

+++ tests/mockserver.py
@@ -59,7 +59,7 @@ class _RequestHandler(BaseHTTPRequestHandler):
         parsed_path = urlparse(self.path)
         query_string = {key: values[0] for key, values in parse_qs(parsed_path.query).items()}
 
-        if delay := int(query_string.get("delay") or 0):
+        if delay := float(query_string.get("delay") or 0):
             print(f"Sleeping {delay} seconds on path {parsed_path.path}...")
             self.server._stop_event.wait(delay)  # type: ignore[attr-defined]
